+++ ui/fetchstrategies.go
@@ -23,7 +23,7 @@ func (client *UI) fetchFromLocalFile() {
 		if reader == nil {
 			return
 		}
-		err = fetchMessage("file", reader.URI().Path())
+		err = fetchMessage("file", reader.URI().Path(), false)
 		if err != nil {
 			dg := dialog.NewError(err, client.window)
 			dg.Show()
@@ -54,7 +54,7 @@ func (client *UI) fetchFromLocalFolder() {
 		progress := dialog.NewCustom("Fetching", "Cancel fetch", hbox, client.window)
 		// TODO: allow fetch to be cancelled by passing a chan
 		progress.Show()
-		err = fetchMessage("directory", file.Path())
+		err = fetchMessage("directory", file.Path(), false)
 		if err != nil {
 			dg := dialog.NewError(err, client.window)
 			dg.Show()
@@ -68,7 +68,7 @@ func (client *UI) fetchFromLocalFolder() {
 	fd.Show()
 }
 
-func (client *UI) fetchFromArchiveServer() {
+func (client *UI) fetchFromArchiveServer(anonOnly bool) {
 	content := container.NewVBox()
 	selectedFetchArchive := widget.NewEntry()
 	content.Add(widget.NewLabel("Press fetch to pull messages from the URL below"))
@@ -84,7 +84,7 @@ func (client *UI) fetchFromArchiveServer() {
 		hbox.Add(pb)
 		progress := dialog.NewCustom("Fetching", "Cancel fetch", hbox, client.window)
 		progress.Show()
-		err := fetchMessage("remoteURL", selectedFetchArchive.Text)
+		err := fetchMessage("remoteURL", selectedFetchArchive.Text, anonOnly)
 		if err != nil {
 			progress.Hide()
 			de := dialog.NewError(err, client.window)
@@ -98,7 +98,7 @@ func (client *UI) fetchFromArchiveServer() {
 	cc.Show()
 }
 
-func fetchMessage(selectedFetchMethod string, location string) (err error) {
+func fetchMessage(selectedFetchMethod string, location string, anonOnly bool) (err error) {
 	switch selectedFetchMethod {
 	case "file":
 		f := fetcher.Fetcher{}
@@ -113,12 +113,17 @@ func fetchMessage(selectedFetchMethod string, location string) (err error) {
 			return err
 		}
 	case "remoteURL":
+		var f *fetcher.Fetcher
 		dir, err := ioutil.TempDir("", "syndie")
 		if err != nil {
 			log.Fatalf("Unable to create a temporary directory: %s", err)
 		}
 		defer os.RemoveAll(dir)
-		f := fetcher.New(location, dir, 60, 50)
+		if anonOnly {
+			f = fetcher.NewAnonOnly(location, dir, 60, 50)
+		} else {
+			f = fetcher.New(location, dir, 60, 50)
+		}
 		err = f.RemoteFetch()
 		if err != nil {
 			return err